wallet: move deprecated PSBT methods into `deprecated.go`
